char oriented frontpage
