+++ src/Pages/DB/index.ts
@@ -1,2 +1,3 @@
-export * from "./DB";
+export { DB } from "./DB";
+export { DbChar } from "./DbChar";
 export * from "./dbSlice";
